Add a new vim-plugin as a python linter
We add python-mode for vim, this can show error info for python files.
I disable most functions of it which I can use ycm to substitute.
Adding this plugin leads to update vim-gitgutter, because the error
signs are below git status signs. In order to resolve this, I changed
the priority of git status signs to 9 to make sure the error signs will
appear on the git status signs.