Update some coc configure and fix installer
`installer.sh` will fail in `ubuntu-20.04`, we find that we place the
`${SUDO_COMMAND}` wrong place, now we fix this.

We now use `<CR>` and `<C-y>` to confirm a selection. We now use `<C-c>`
to cancel a completion. Because now we can use `<CR>` to confirm a
selection, so we now update our configure to select the first item by
default.