Change the theme and complete more configurations
We find the old theme works not well with the coc's go references, so we
just choose `vimone` as our new theme. Besides, we find that `<CR>` can
not confirm a completion in a `vimwiki` file, so we remap the `<CR>` in
`vimwiki` to fix this. Besides, we add `<buffer>` at the
`markdownhelper_config` so that the auto commands will only be efficient
for the current buffer rather than when returning a non-markdown file.